+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 
 ### Bug fixes
 
+- [#472][] Clear terminal title notification on exit. Fixes [#472][]. ([@netzpirat][])
 - [#457][] Raise an exception when a group is called "all". Fixes [#457][]. (reported by [@rweng][], fixed by [@rymai][])
 - [#471][] Only init once per plugin when running `guard init` (reported by [@simon-ohara][], fixed by [@thibaudgg][])
 - [#456][] Fix notifu notifier. ([@netzpirat][])
@@ -839,6 +840,7 @@ The Listen integration has been supervised by [@thibaudgg][] and executed by [@M
 [#460]: https://github.com/guard/guard/issues/460
 [#463]: https://github.com/guard/guard/issues/463
 [#471]: https://github.com/guard/guard/issues/471
+[#472]: https://github.com/guard/guard/issues/472
 [@Gazer]: https://github.com/Gazer
 [@Maher4Ever]: https://github.com/Maher4Ever
 [@Nerian]: https://github.com/Nerian